cleanup: Remove old Otlp Class
This commit removes the old Otlp Exporter which wasn't to spec. I feel it's causing confusion about it being there.

On a side note, I was going to put common parts of OtlpGrpc and OtlpHttp into this namespace but haven't got around to it yet. So small clean up commit, until I find the time.