Add integration tests for tracing compliance (#662)
* Add integration tests for span builder

* Space fixed

* Add more descriptive names and @group annotation

Modified the names of the tests to be more descriptive and added the
`@group compliance` notation.
Additionally, I commented out one of the tests that I was not totally
sure of. (unsure if `setAttribute()` should work with empty string `''`
values.

* Fixed test name

* Check for specific attributes and not just count

Removed `$span->end()` calls as they shouldn't be necessary

* Remove redundant `end()` calls

* Add test for getting default tracer

Test to get default tracer. Also, added `@group` php unit annotation for
tests to check specification compliance

* Fixed warning

* Fixed style warning