refactor semantic conventions (#1659)
Align with how the spec advises SemConv should be organised, and break unstable
semconvs out into their own directory.
I played around with generating values as enums, but I could only get this to work
if the output was in the same file as the attributes (which would not work well with
composer autoloading), so I've used <attribute>_VALUE_<name>, similar to what we have
now except all in the same file.

* fix linting, broken values containing '.'

* fix incubating namespace

* update semconv readme

* restore ResourceAttributeValues

* adding DNS semconv

* adding db metrics